+++ src/main/java/chess/command/MoveCommand.java
@@ -1,7 +1,7 @@
 package chess.command;
 
-import chess.action.Action;
-import chess.domain.Status;
+import chess.domain.game.Action;
+import chess.domain.game.Status;
 import chess.domain.position.Position;
 import java.util.List;
