+++ src/main/java/chess/controller/ChessController.java
@@ -2,9 +2,9 @@ package chess.controller;
 
 import chess.command.Command;
 import chess.command.CommandFactory;
-import chess.domain.ChessGame;
-import chess.domain.Status;
 import chess.domain.board.Board;
+import chess.domain.game.ChessGame;
+import chess.domain.game.Status;
 import chess.view.InputView;
 import chess.view.OutputView;
 import java.util.List;
