--- src/main/java/chess/domain/Status.java
@@ -1,5 +1,0 @@
-package chess.domain;
-
-public class Status {
-
-}
