+++ src/main/java/chess/domain/game/Status.java
@@ -1,0 +1,5 @@
+package chess.domain.game;
+
+public class Status {
+
+}
